+++ aristonremotethermo/aristonaqua.py
@@ -67,7 +67,7 @@ class AquaAristonHandler:
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     """
 
-    _VERSION = "1.0.32"
+    _VERSION = "1.0.33"
 
     _LOGGER = logging.getLogger(__name__)
     _LEVEL_CRITICAL = "CRITICAL"
@@ -706,6 +706,11 @@ class AquaAristonHandler:
           - supported parameters to be changed by API (actual list of supported parameters by the model cannot be identified and must be chosen manually)
         """
         return cls._VERSION, cls._SENSOR_LIST, cls._SENSOR_SET_LIST
+
+    @property
+    def plant_id(self) -> str:
+        """Return the unique plant_id."""
+        return self._plant_id
 
     @property
     def available(self) -> bool: